Proposal for a more compact `miss` encoding.
Instead of sending the list of full missing `seq` values.
It seems more optimal to send a list of delta encoded `seq` ids
based on the accompanying `ack` (which must always be smaller).

For example:

Instead of sending:

    {
      "ack": 78231,
      "miss": [78236, 78235, 78245, 78238]
    }

We could send:

    {
      "ack": 78231,
      "miss": [4, 1, 2, 7]
    }

This should significantly reduce the impact of reliable channels.
Especially over very unreliable transports with small MTUs.